Dashing_board/requirements.txt
Django==2.1.2
django-dashing==0.3.3
pip==10.0.1
pytz==2018.7
setuptools==39.0.1